Clean up and url bug fix
