+++ app/Models/User.php
@@ -5,58 +5,36 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
-use App\Models\Shift;
-use App\Models\Project;
 
 class User extends Authenticatable
 {
-    /** @use HasFactory<\Database\Factories\UserFactory> */
     use HasFactory, Notifiable;
 
-
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var list<string>
-     */
     protected $fillable = [
         'name',
         'netid',
         'email',
-        'password',
         'active',
-        'is_admin'
     ];
 
-    /**
-     * The attributes that should be hidden for serialization.
-     *
-     * @var list<string>
-     */
+    protected $guarded = [
+        'is_admin',
+    ];
+
     protected $hidden = [
         'remember_token',
     ];
-    
-    protected $primaryKey = 'id';
-    public $incrementing = false;
-    protected $keyType = 'string';
 
-    /**
-     * Get the attributes that should be cast.
-     *
-     * @return array<string, string>
-     */
     protected function casts(): array
     {
         return [
             'email_verified_at' => 'datetime',
-            'password' => 'hashed',
             'active' => 'boolean',
             'is_admin' => 'boolean',
         ];
     }
 
-    public function isAdmin()
+    public function isAdmin(): bool
     {
         return $this->is_admin;
     }
@@ -64,43 +42,12 @@ class User extends Authenticatable
     public function projects()
     {
         return $this->belongsToMany(Project::class, 'project_user', 'user_netid', 'project_id', 'netid', 'id')
-                    ->withPivot('active')
-                    ->withTimestamps();
+            ->withPivot('active')
+            ->withTimestamps();
     }
 
     public function shifts()
     {
         return $this->hasMany(Shift::class, 'netid', 'netid');
-    }
-
-    public function scopeSearch($query, ?string $search)
-    {
-        if (!$search) {
-            return $query;
-        }
-
-        return $query->where(function ($q) use ($search) {
-            $q->where('name', 'like', '%' . $search . '%')
-              ->orWhere('netid', 'like', '%' . $search . '%')
-              ->orWhere('email', 'like', '%' . $search . '%');
-        });
-    }
-
-    public function scopeFilterAdmin($query, $adminFilter)
-    {
-        if ($adminFilter === null || $adminFilter === '') {
-            return $query;
-        }
-
-        return $query->where('is_admin', $adminFilter == '1');
-    }
-
-    public function scopeFilterActive($query, $activeFilter)
-    {
-        if ($activeFilter === null || $activeFilter === '') {
-            return $query;
-        }
-
-        return $query->where('active', $activeFilter == '1');
     }
 }
